Replace booster() with get_booster()
The booster member in XGBClassifier is demoted from a function to a string variable in XGB version since CMSSW 102X. The proper way to access this information is via get_booster() member function.
This is one of the reasons why the training doesn't work in CMSSW 102X or later.
Related: https://github.com/HEP-KBFI/tth-htt/issues/73